上游这轮大重构(Handlers -> Flows, 拆主界面 VM), 冲突 8 处, 处理原则: 我们的功能钩子照留, 排版跟上游新风格。

- GameRunner.cs 被上游删除, 逻辑内联进 GameLaunchFlow -> 接受删除。
  防游戏随机闪退的关键在 GameArgumentInterop.KeepForever(GC.SuppressFinalize), 那个文件干净合并, 修复仍在。
- GameLaunchFlow: 保留 InGameAgents 判定, 改用上游的 dotnetRuntimePath。
- DCTravelRuntimeService / LoginFlow: 保留无参 StartAsync(F2 定案: 不按注入代理卡会话)。
- MainWindow.xaml.cs: headlines 等字段上游已挪进 NewsFlow, 只留注入列的宽度常量; 两边的事件订阅/方法都留。
- MainWindowViewModel: RefreshDalamudInfoCommandState -> DalamudStatus.RefreshCommandState, 注入选项回填照留。
- DCTravelViewModel / SettingsWindowViewModel: 两边都留。

Debug 构建 0 错误; 未实机跑。

+++ src/XIVLauncher.GamePatchV3/Update/Models/GamePatchProgress.cs
@@ -5,8 +5,8 @@ public sealed class GamePatchProgress
     public string PhaseText      { get; init; } = string.Empty;
     public string CurrentFile    { get; init; } = string.Empty;
     public string StatusText     { get; init; } = string.Empty;
-    public long   Progress       { get; init; }
-    public long   Total          { get; init; }
+    public double Progress       { get; init; }
+    public double Total          { get; init; }
     public long   Speed          { get; init; }
     public bool   IsByteProgress { get; init; }
 }
